Develop justin

+++ src/main/java/com/whydigit/efit/repo/AssetRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.whydigit.efit.entity.AssetVO;
 
 @Repository
-public interface AssetRepo extends JpaRepository<AssetVO, Integer> {
+public interface AssetRepo extends JpaRepository<AssetVO, Long> {
 	@Query(value = "select a from AssetVO a Where a.orgId=?1")
 	List<AssetVO> getAllAssetByOrgId(Long orgId);
